File move from original directory to edit directory- web-dev-hw
I lost my original web-dev-hw directory, and created an edit directory to migrate work. I'm not sure if this was a proper work around- but, it seems to be addressing the issue that I was having. I will likely delete the original and rename the edit. I still haven't decided.