Assignment 3 push-1
There are the changes that I have made for assignment 3. I am still experiencing difficulties with uploading images, and links. I think I might be able to troubleshoot this issue if I were able to view the interactive images. I'm still trying to figure out how to mend this issue.